requirements.txt
# Core
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-Cors==4.0.0
python-dotenv==1.0.0

# Database
pymongo==4.5.0
Flask-PyMongo==2.3.0

# Authentication
Flask-Login==0.6.2
Flask-JWT-Extended==4.5.2
flask-bcrypt==1.0.1
Authlib==1.2.1
python-jose==3.3.0

# Payments
stripe==5.5.0

# AI/ML
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.0
protobuf==3.20.3
plotly==5.15.0
joblib==1.3.2
transformers
torch

# Utilities
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
Werkzeug==2.3.8
google-api-python-client>=2.0.0
google-auth-oauthlib>=0.4.1.7
gunicorn==21.2.0

# Development
pytest==7.4.0
black==23.7.0
flake8==6.1.0
